Move method and staticMethod matching to src/javascript from src/polymer
