+++ src/polymer/js-utils.ts
@@ -13,12 +13,11 @@
  */
 
 import * as doctrine from 'doctrine';
-import * as escodegen from 'escodegen';
 import * as estree from 'estree';
 
-import {closureType, getAttachedComment, objectKeyToString} from '../javascript/esutil';
+import {closureType, configurationProperties, getAttachedComment, getOrInferPrivacy, objectKeyToString} from '../javascript/esutil';
 import * as jsdoc from '../javascript/jsdoc';
-import {Privacy, ScannedMethod, Severity, SourceRange, Warning} from '../model/model';
+import {Severity, SourceRange, Warning} from '../model/model';
 import {ParsedDocument} from '../parser/document';
 
 import {ScannedPolymerProperty} from './polymer-element';
@@ -66,104 +65,3 @@ export function toScannedPolymerProperty(
 
   return result;
 };
-
-/**
- * Properties on Polymer element prototypes that are part of Polymer's
- * configuration syntax.
- */
-const configurationProperties = new Set([
-  'attached',
-  'attributeChanged',
-  'beforeRegister',
-  'configure',
-  'constructor',
-  'created',
-  'detached',
-  'enableCustomStyleProperties',
-  'extends',
-  'hostAttributes',
-  'is',
-  'listeners',
-  'mixins',
-  'observers',
-  'properties',
-  'ready',
-  'registered',
-]);
-
-/**
- * Create a ScannedMethod object from an estree Property AST node.
- */
-export function toScannedMethod(
-    node: estree.Property|estree.MethodDefinition,
-    sourceRange: SourceRange,
-    document: ParsedDocument<any, any>): ScannedMethod {
-  const scannedMethod: ScannedMethod =
-      toScannedPolymerProperty(node, sourceRange, document);
-
-  if (scannedMethod.type === 'Function' ||
-      scannedMethod.type === 'ArrowFunction') {
-    const value = <estree.FunctionExpression>node.value;
-
-    const paramTags = new Map<string, doctrine.Tag>();
-    if (scannedMethod.jsdoc) {
-      for (const tag of (scannedMethod.jsdoc.tags || [])) {
-        if (tag.title === 'param' && tag.name) {
-          paramTags.set(tag.name, tag);
-
-        } else if (tag.title === 'return' || tag.title === 'returns') {
-          scannedMethod.return = {};
-          if (tag.type) {
-            scannedMethod.return.type = doctrine.type.stringify(tag.type!);
-          }
-          if (tag.description) {
-            scannedMethod.return.desc = tag.description;
-          }
-        }
-      }
-    }
-
-    scannedMethod.params = (value.params || []).map((nodeParam) => {
-      let type = undefined;
-      let description = undefined;
-      // With ES6 we can have a lot of param patterns. Best to leave the
-      // formatting to escodegen.
-      const name = escodegen.generate(nodeParam);
-      const tag = paramTags.get(name);
-      if (tag) {
-        if (tag.type) {
-          type = doctrine.type.stringify(tag.type);
-        }
-        if (tag.description) {
-          description = tag.description;
-        }
-      }
-      return {name, type, description};
-    });
-  }
-
-  return scannedMethod;
-}
-
-
-export function getOrInferPrivacy(
-    name: string,
-    annotation: jsdoc.Annotation|undefined,
-    defaultPrivacy: Privacy = 'public'): Privacy {
-  const explicitPrivacy = jsdoc.getPrivacy(annotation);
-  const specificName = name.slice(name.lastIndexOf('.') + 1);
-
-  if (explicitPrivacy) {
-    return explicitPrivacy;
-  }
-  if (specificName.startsWith('__')) {
-    return 'private';
-  } else if (specificName.startsWith('_')) {
-    return 'protected';
-  } else if (specificName.endsWith('_')) {
-    return 'private';
-  } else if (configurationProperties.has(specificName)) {
-    return 'protected';
-  }
-  return defaultPrivacy;
-}
